Rename Car views and controllers
CarConfigurationView = EditCarView
NewCarCreationView = NewCarView
