+++ src/main/java/de/htwsaar/prog3/carrental/util/i18n/I18nStringsUtil.java
@@ -10,8 +10,8 @@ import java.util.ResourceBundle;
 public class I18nStringsUtil {
     // Keys in strings_*.properties
     private static final String CAR_TABLE_VIEW_URL = "car-tableview-fxml-file";
-    private static final String NEW_CAR_CREATION_VIEW_URL = "new-car-creationview-fxml-file";
-    private static final String CAR_CONFIGURATION_VIEW_IRL = "car-configurationview-fxml-file";
+    private static final String NEW_CAR_VIEW_URL = "new-carview-fxml-file";
+    private static final String EDIT_CAR_VIEW_URL = "edit-carview-fxml-file";
     private static final String EMPLOYEE_TABLE_VIEW_URL = "employee-tableview-fxml-file";
     private static final String RENTAL_TABLE_VIEW_URL = "rental-tableview-fxml-file";
     private static final String CUSTOMER_TABLE_VIEW_URL = "customer-tableview-fxml-file";
@@ -31,12 +31,12 @@ public class I18nStringsUtil {
         return resourceBundle.getString(CAR_TABLE_VIEW_URL);
     }
 
-    public static String getNewCarCreationViewURL() {
-        return resourceBundle.getString(NEW_CAR_CREATION_VIEW_URL);
+    public static String getNewCarViewURL() {
+        return resourceBundle.getString(NEW_CAR_VIEW_URL);
     }
 
-    public static String getCarConfigurationViewURL() {
-        return resourceBundle.getString(CAR_CONFIGURATION_VIEW_IRL);
+    public static String getEditCarViewURL() {
+        return resourceBundle.getString(EDIT_CAR_VIEW_URL);
     }
 
     public static String getEmployeeTableViewURL() {
