+++ src/main/resources/bundles/strings.properties
@@ -1,0 +1,13 @@
+car-tableview-fxml-file=/fxml/CarTableView.fxml
+new-carview-fxml-file=/fxml/NewCarView.fxml
+edit-carview-fxml-file=/fxml/EditCarView.fxml
+
+customer-tableview-fxml-file=/fxml/CustomerTableView.fxml
+new-customerview-fxml-file=/fxml/NewCustomerView.fxml
+edit-customerview-fxml-file=/fxml/EditCustomerView.fxml
+
+employee-tableview-fxml-file=/fxml/EmployeeTableView.fxml
+new-employeeview-fxml-file=/fxml/NewEmployeeView.fxml
+edit-employeeview-fxml-file=/fxml/EditEmployeeView.fxml
+
+rental-tableview-fxml-file=/fxml/RentalTableView.fxml
